% Improved Design Quality Used Fragments + Added Splash Icon
