+++ app/src/main/java/com/example/route_task/features/products/ProductsViewModel.kt
@@ -22,7 +22,6 @@ import javax.inject.Inject
 @HiltViewModel
 class ProductsViewModel @Inject constructor(
     private val getProductsUseCase: GetProductsUseCase
-
 ) : ViewModel() {
     val shouldShowLoading = MutableLiveData<Boolean>()
     val shouldShowError = MutableLiveData<Boolean>()
