+++ app/src/main/res/layout/header_search_cart.xml
@@ -3,24 +3,12 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:layout_marginTop="0dp"
     android:layout_weight="9"
-    android:paddingHorizontal="10dp">
+    android:paddingHorizontal="8dp">
 
-    <ImageView
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintBottom_toTopOf="@id/searchBar"
-
-        android:layout_marginStart="8dp"
-        android:id="@+id/appbar"
-        android:layout_width="66dp"
-        android:layout_height="22dp"
-        android:contentDescription="@string/logo"
-        android:src="@drawable/ic_logo" />
 
     <EditText
-        app:layout_constraintTop_toBottomOf="@id/appbar"
+        app:layout_constraintTop_toTopOf="parent"
         android:id="@+id/searchBar"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
@@ -46,10 +34,10 @@
         android:layout_gravity="center"
         android:layout_marginEnd="16dp"
         android:layout_marginTop="16dp"
+        app:layout_constraintTop_toTopOf="parent"
         android:layout_weight="1"
         android:contentDescription="@string/cartIcon"
         app:layout_constraintStart_toEndOf="@id/searchBar"
         android:src="@drawable/ic_cart"
-        app:layout_constraintEnd_toEndOf="parent"
-       app:layout_constraintTop_toBottomOf="@id/appbar" />
+        app:layout_constraintEnd_toEndOf="parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
