+++ app/src/main/res/layout/item_product.xml
@@ -39,11 +39,11 @@
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp"
                 android:layout_marginStart="8dp"
-                android:fontFamily="@font/poppins_semibold"
+                android:fontFamily="@font/poppins_medium"
                 android:lines="1"
                 android:maxLines="1"
                 android:text="@{product.title}"
-                android:textColor="@color/desc_dark_blue"
+                android:textColor="@color/main_text_color"
                 android:textSize="14sp"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
@@ -54,12 +54,12 @@
                 android:layout_height="wrap_content"
                 android:layout_marginBottom="8dp"
                 android:layout_marginStart="8dp"
-                android:fontFamily="@font/poppins_semibold"
+                android:fontFamily="@font/poppins_medium"
                 android:lines="1"
                 android:maxLines="1"
                 android:ellipsize="end"
                 android:text="@{product.description}"
-                android:textColor="@color/desc_dark_blue"
+                android:textColor="@color/main_text_color"
                 android:textSize="14sp"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
@@ -94,9 +94,9 @@
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp"
-                android:fontFamily="@font/poppins_semibold"
+                android:fontFamily="@font/poppins_medium"
                 android:text="@string/egp_1_200"
-                android:textColor="@color/desc_dark_blue"
+                android:textColor="@color/main_text_color"
                 android:textSize="14sp"
                 app:layout_constraintStart_toStartOf="@id/product_desc"
                 app:layout_constraintTop_toBottomOf="@id/product_desc" />
@@ -106,7 +106,7 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="10dp"
-                android:fontFamily="@font/poppins_semibold"
+                android:fontFamily="@font/poppins_medium"
                 android:text="@string/egp_1_500"
                 android:textColor="@color/pale_blue"
                 android:textSize="11sp"
@@ -119,9 +119,9 @@
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="4dp"
-                android:fontFamily="@font/poppins_semibold"
+                android:fontFamily="@font/poppins_medium"
                 android:text="@string/review"
-                android:textColor="@color/desc_dark_blue"
+                android:textColor="@color/main_text_color"
                 android:textSize="14sp"
                 app:layout_constraintStart_toStartOf="@id/product_price"
                 app:layout_constraintTop_toBottomOf="@id/product_price" />
@@ -132,8 +132,8 @@
                 android:layout_height="wrap_content"
                 android:layout_marginStart="5dp"
                 android:layout_marginTop="4dp"
-                android:fontFamily="@font/poppins_semibold"
-                android:textColor="@color/desc_dark_blue"
+                android:fontFamily="@font/poppins_medium"
+                android:textColor="@color/main_text_color"
                 android:textSize="14sp"
                 app:layout_constraintStart_toEndOf="@id/review_tv"
                 app:layout_constraintTop_toBottomOf="@id/product_price"
