+++ app/src/main/res/values/strings.xml
@@ -15,4 +15,6 @@
     <string name="try_again">Try Again</string>
     <string name="review">Review</string>
     <string name="star">star</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
