+++ domain/src/main/java/com/example/domain/products/usecase/GetProductsUseCase.kt
@@ -12,6 +12,4 @@ class GetProductsUseCase @Inject constructor(
     suspend fun invoke(): ProductsEntity {
         return repository.getProducts()
     }
-
-
 }
